apps/merchy_backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
django==4.1.1
django-allauth==0.43.0
django-cors-headers==3.0.2
django-environ==0.4.5
django-otp==0.9.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
dj-rest-auth==1.0.0
djangorestframework==3.14.0
djoser==2.0.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.5.4
urllib3==1.25.9
wrapt==1.11.2
